remove request package and add funcs to handler
